rename float-packing to number-packing
- more telling name
