+++ src/mol-util/number-packing.ts
